+++ cli/commands/index.ts
@@ -1,8 +1,10 @@
 import { Command } from '@oclif/core';
 
+import ClearCache from './clear-cache';
 import CodegenApp from './codegen-app';
 import CreateAppCommand from './create-app';
 import GenerateContractFactory from './generate-contract-factory';
+import SetNetworkProvider from './set-network-provider';
 
 /**
  * Record of invocation string and the associated command.
@@ -13,4 +15,6 @@ export const commands: Record<string, typeof Command> = {
   'generate:contract-factory': GenerateContractFactory,
   'create-app': CreateAppCommand,
   'codegen-app': CodegenApp,
+  'clear-cache': ClearCache,
+  'set-network-provider': SetNetworkProvider,
 };