+++ src/cache/cache.service.ts
@@ -1,6 +1,6 @@
-import { CACHE_MANAGER, Inject, Injectable, Logger, OnModuleInit } from '@nestjs/common';
+import { Inject, Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { DiscoveryService, MetadataScanner, Reflector } from '@nestjs/core';
-import { Cache } from 'cache-manager';
+import Cache from 'file-system-cache';
 import { isNil } from 'lodash';
 
 import { CacheOptions, CACHE_KEY, CACHE_TTL } from './cache.decorator';
@@ -8,9 +8,12 @@ import { CacheOptions, CACHE_KEY, CACHE_TTL } from './cache.decorator';
 @Injectable()
 export class CacheService implements OnModuleInit {
   private logger = new Logger(CacheService.name);
+  private cacheManager = Cache({
+    basePath: './.cache',
+    ns: '@Cache',
+  });
 
   constructor(
-    @Inject(CACHE_MANAGER) private readonly cacheManager: Cache,
     @Inject(DiscoveryService) private readonly discoveryService: DiscoveryService,
     @Inject(MetadataScanner) private readonly metadataScanner: MetadataScanner,
     @Inject(Reflector) private readonly reflector: Reflector,